# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Service Assurance AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: string
        - name: siteId
          in: query
          schema:
            type: string
        - name: vendorName
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketWithRelations"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTicket
      tags: [tickets]
      summary: Update a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/updates:
    get:
      operationId: getTicketUpdates
      tags: [ticket-updates]
      summary: Get updates for a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketUpdate"
    post:
      operationId: createTicketUpdate
      tags: [ticket-updates]
      summary: Add an update to a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketUpdateRequest"
      responses:
        "201":
          description: Created ticket update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketUpdate"
  /tickets/{id}/ai/summarize:
    post:
      operationId: aiSummarizeTicket
      tags: [ai]
      summary: Generate AI summary for a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSummarizeResponse"
  /tickets/{id}/ai/normalize-latest-update:
    post:
      operationId: aiNormalizeLatestUpdate
      tags: [ai]
      summary: Normalize the latest vendor update status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI normalized status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiNormalizeResponse"
  /tickets/{id}/ai/generate-customer-update:
    post:
      operationId: aiGenerateCustomerUpdate
      tags: [ai]
      summary: Generate a customer-facing update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI customer update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCustomerUpdateResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, ops, customer]
        customerId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive]
        primaryContactName:
          type: string
          nullable: true
        primaryContactEmail:
          type: string
          nullable: true
        primaryContactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
...
    Site:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteName:
          type: string
        address1:
          type: string
          nullable: true
        address2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        siteCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteName
        - createdAt
        - updatedAt
...
    Service:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        siteId:
          type: string
        vendorName:
          type: string
        serviceType:
          type: string
          enum: [DIA, Broadband, SD-WAN, Voice, Wireless, Other]
        circuitId:
          type: string
          nullable: true
        bandwidth:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, pending, down, impaired, disconnected]
        installDate:
          type: string
          nullable: true
        monthlyRecurringCharge:
          type: number
          nullable: true
        supportReference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - siteId
        - vendorName
        - serviceType
        - status
        - createdAt
        - updatedAt
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        customerId:
          type: string
        siteId:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, email, api]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [new, investigating, vendor_engaged, dispatch_scheduled, monitoring, resolved, closed]
        outageType:
          type: string
          enum: [outage, impairment, informational, unknown]
        vendorTicketId:
          type: string
          nullable: true
        assignedToUserId:
          type: string
          nullable: true
        openedAt:
          type: string
          format: date-time
        lastUpdatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        nextEscalationAt:
          type: string
          format: date-time
          nullable: true
        slaTargetMinutes:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        aiNormalizedStatus:
          type: string
          nullable: true
        aiCustomerUpdate:
          type: string
          nullable: true
        aiLastGeneratedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticketNumber
        - customerId
        - title
        - source
        - severity
        - status
        - outageType
        - openedAt
        - lastUpdatedAt
        - createdAt
        - updatedAt
    TicketWithRelations:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            customer:
              $ref: "#/components/schemas/Customer"
            site:
              $ref: "#/components/schemas/Site"
            service:
              $ref: "#/components/schemas/Service"
            assignedTo:
              $ref: "#/components/schemas/User"
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/TicketWithRelations"
        - type: object
          properties:
            updates:
              type: array
              items:
                $ref: "#/components/schemas/TicketUpdate"
    CreateTicketRequest:
      type: object
      properties:
        customerId:
          type: string
        siteId:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        source:
          type: string
          enum: [manual, email, api]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [new, investigating, vendor_engaged, dispatch_scheduled, monitoring, resolved, closed]
        outageType:
          type: string
          enum: [outage, impairment, informational, unknown]
        vendorTicketId:
          type: string
          nullable: true
        assignedToUserId:
          type: string
          nullable: true
        slaTargetMinutes:
          type: number
          nullable: true
      required:
        - customerId
        - title
        - source
        - severity
        - status
        - outageType
    UpdateTicketRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [new, investigating, vendor_engaged, dispatch_scheduled, monitoring, resolved, closed]
        outageType:
          type: string
          enum: [outage, impairment, informational, unknown]
        vendorTicketId:
          type: string
          nullable: true
        assignedToUserId:
          type: string
          nullable: true
        nextEscalationAt:
          type: string
          format: date-time
          nullable: true
        slaTargetMinutes:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        aiNormalizedStatus:
          type: string
          nullable: true
        aiCustomerUpdate:
          type: string
          nullable: true
    TicketUpdate:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        updateType:
          type: string
          enum: [internal_note, vendor_update, customer_update, system_event, ai_generated]
        rawText:
          type: string
        normalizedStatus:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [internal, customer]
        createdByUserId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: "#/components/schemas/User"
          nullable: true
      required:
        - id
        - ticketId
        - updateType
        - rawText
        - visibility
        - createdAt
    CreateTicketUpdateRequest:
      type: object
      properties:
        updateType:
          type: string
          enum: [internal_note, vendor_update, customer_update, system_event, ai_generated]
        rawText:
          type: string
        normalizedStatus:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [internal, customer]
      required:
        - updateType
        - rawText
        - visibility
...
    AiSummarizeResponse:
      type: object
      properties:
        summary:
          type: string
        ticket:
          $ref: "#/components/schemas/Ticket"
      required:
        - summary
        - ticket
    AiNormalizeResponse:
      type: object
      properties:
        normalizedStatus:
          type: string
        ticket:
          $ref: "#/components/schemas/Ticket"
      required:
        - normalizedStatus
        - ticket
    AiCustomerUpdateResponse:
      type: object
      properties:
        customerUpdate:
          type: string
        ticket:
          $ref: "#/components/schemas/Ticket"
      required:
        - customerUpdate
        - ticket
...